BSc Thesis, Business Administration. Summer 2018.
My Bachelor Thesis on Electrification and Employment in Rural Africa. 
Written for my degree in Business Administration in Summer 2018.

Abstract:
In this Bachelor Thesis, I look at the effects of a grid connection on remote rural households in Rwanda. Most of these households never before had any electricity. I find that both electricity and appliance ownership (independent from each other) are negatively correlated  with how remote a given household is. This holds true after controlling for other investments and spending levels in general. Therefore, remote communities are underserved and would benefit from  improved contact with the outside world.
The literature expects the gender gap in spending and time worked to be reduced by electricity. This does not seem to be the case in an environment with little-developed electricitymarkets. Women reap no additional benefits from electricity. 
I find that small businesses are particularly quick to adopt electricity with higher usage and appliance ownership rates. Spending does not increase. It seems, however, as if they are saving more than the average household. They have both lower spending on micro-credits and higher rates of savings accounts.